requirements.txt
alethic-ism-core==1.0.47
alethic-ism-db==1.0.44
pyyaml
nats-py
python-dotenv
pydantic
psycopg2